Serials VFS refactoring
