+++ components/at/src/at_cmd_serial_group.c
@@ -8,7 +8,7 @@ static int vars_serial_read(const struct cat_variable* var, const struct cat_com
 {
     serial_id_t id = cmd->name[strlen(cmd->name)] - '0';
 
-    var_u8_1 = serial_get_mode(id);
+    strcpy(var_str32_1,  serial_get_mode(id));
     var_i32_1 = serial_get_baud_rate(id);
     var_u8_2 = serial_get_data_bits(id);
     var_u8_3 = serial_get_stop_bits(id);
@@ -20,7 +20,7 @@ static int vars_serial_read(const struct cat_variable* var, const struct cat_com
 static cat_return_state vars_serial_write(const struct cat_command* cmd, const uint8_t* data, const size_t data_size, const size_t args_num)
 {
     serial_id_t id = cmd->name[strlen(cmd->name)] - '0';
-    serial_mode_t mode = var_u8_1;
+    const char* mode = var_str32_1;
     int baud_rate = var_i32_1;
     uart_word_length_t data_bits = var_u8_2;
     uart_stop_bits_t stop_bits = var_u8_3;
@@ -31,9 +31,9 @@ static cat_return_state vars_serial_write(const struct cat_command* cmd, const u
 
 static struct cat_variable vars_serial[] = {
     {
-        .type = CAT_VAR_UINT_DEC,
-        .data = &var_u8_1,
-        .data_size = sizeof(var_u8_1),
+        .type = CAT_VAR_BUF_STRING,
+        .data = &var_str32_1,
+        .data_size = sizeof(var_str32_1),
         .access = CAT_VAR_ACCESS_READ_WRITE,
         .read_ex = vars_serial_read,
     },
