+++ components/logger/CMakeLists.txt
@@ -1,2 +1,4 @@
 idf_component_register(SRC_DIRS "src"
-                    INCLUDE_DIRS "include")
+                    INCLUDE_DIRS "include"
+                    PRIV_REQUIRES app_update
+                    REQUIRES network config)
