--- components/serial/src/serial_modbus.c
@@ -1,167 +1,0 @@
-#include "serial_modbus.h"
-
-#include <esp_log.h>
-
-#include "modbus.h"
-
-#define BUF_SIZE         256
-#define EVENT_QUEUE_SIZE 20
-
-#define LOG_LVL_DATA ESP_LOG_VERBOSE
-
-static const char* TAG = "serial_modbus";
-
-static uart_port_t port = -1;
-
-static TaskHandle_t serial_modbus_task = NULL;
-
-static QueueHandle_t uart_queue;
-
-static const uint8_t crc_hi[] = {
-    0x00, 0xC1, 0x81, 0x40, 0x01, 0xC0, 0x80, 0x41, 0x01, 0xC0, 0x80, 0x41, 0x00, 0xC1, 0x81, 0x40, 0x01, 0xC0, 0x80, 0x41, 0x00, 0xC1, 0x81, 0x40, 0x00, 0xC1, 0x81, 0x40, 0x01,
-    0xC0, 0x80, 0x41, 0x01, 0xC0, 0x80, 0x41, 0x00, 0xC1, 0x81, 0x40, 0x00, 0xC1, 0x81, 0x40, 0x01, 0xC0, 0x80, 0x41, 0x00, 0xC1, 0x81, 0x40, 0x01, 0xC0, 0x80, 0x41, 0x01, 0xC0,
-    0x80, 0x41, 0x00, 0xC1, 0x81, 0x40, 0x01, 0xC0, 0x80, 0x41, 0x00, 0xC1, 0x81, 0x40, 0x00, 0xC1, 0x81, 0x40, 0x01, 0xC0, 0x80, 0x41, 0x00, 0xC1, 0x81, 0x40, 0x01, 0xC0, 0x80,
-    0x41, 0x01, 0xC0, 0x80, 0x41, 0x00, 0xC1, 0x81, 0x40, 0x00, 0xC1, 0x81, 0x40, 0x01, 0xC0, 0x80, 0x41, 0x01, 0xC0, 0x80, 0x41, 0x00, 0xC1, 0x81, 0x40, 0x01, 0xC0, 0x80, 0x41,
-    0x00, 0xC1, 0x81, 0x40, 0x00, 0xC1, 0x81, 0x40, 0x01, 0xC0, 0x80, 0x41, 0x01, 0xC0, 0x80, 0x41, 0x00, 0xC1, 0x81, 0x40, 0x00, 0xC1, 0x81, 0x40, 0x01, 0xC0, 0x80, 0x41, 0x00,
-    0xC1, 0x81, 0x40, 0x01, 0xC0, 0x80, 0x41, 0x01, 0xC0, 0x80, 0x41, 0x00, 0xC1, 0x81, 0x40, 0x00, 0xC1, 0x81, 0x40, 0x01, 0xC0, 0x80, 0x41, 0x01, 0xC0, 0x80, 0x41, 0x00, 0xC1,
-    0x81, 0x40, 0x01, 0xC0, 0x80, 0x41, 0x00, 0xC1, 0x81, 0x40, 0x00, 0xC1, 0x81, 0x40, 0x01, 0xC0, 0x80, 0x41, 0x00, 0xC1, 0x81, 0x40, 0x01, 0xC0, 0x80, 0x41, 0x01, 0xC0, 0x80,
-    0x41, 0x00, 0xC1, 0x81, 0x40, 0x01, 0xC0, 0x80, 0x41, 0x00, 0xC1, 0x81, 0x40, 0x00, 0xC1, 0x81, 0x40, 0x01, 0xC0, 0x80, 0x41, 0x01, 0xC0, 0x80, 0x41, 0x00, 0xC1, 0x81, 0x40,
-    0x00, 0xC1, 0x81, 0x40, 0x01, 0xC0, 0x80, 0x41, 0x00, 0xC1, 0x81, 0x40, 0x01, 0xC0, 0x80, 0x41, 0x01, 0xC0, 0x80, 0x41, 0x00, 0xC1, 0x81, 0x40,
-};
-
-static const uint8_t crc_lo[] = {
-    0x00, 0xC0, 0xC1, 0x01, 0xC3, 0x03, 0x02, 0xC2, 0xC6, 0x06, 0x07, 0xC7, 0x05, 0xC5, 0xC4, 0x04, 0xCC, 0x0C, 0x0D, 0xCD, 0x0F, 0xCF, 0xCE, 0x0E, 0x0A, 0xCA, 0xCB, 0x0B, 0xC9,
-    0x09, 0x08, 0xC8, 0xD8, 0x18, 0x19, 0xD9, 0x1B, 0xDB, 0xDA, 0x1A, 0x1E, 0xDE, 0xDF, 0x1F, 0xDD, 0x1D, 0x1C, 0xDC, 0x14, 0xD4, 0xD5, 0x15, 0xD7, 0x17, 0x16, 0xD6, 0xD2, 0x12,
-    0x13, 0xD3, 0x11, 0xD1, 0xD0, 0x10, 0xF0, 0x30, 0x31, 0xF1, 0x33, 0xF3, 0xF2, 0x32, 0x36, 0xF6, 0xF7, 0x37, 0xF5, 0x35, 0x34, 0xF4, 0x3C, 0xFC, 0xFD, 0x3D, 0xFF, 0x3F, 0x3E,
-    0xFE, 0xFA, 0x3A, 0x3B, 0xFB, 0x39, 0xF9, 0xF8, 0x38, 0x28, 0xE8, 0xE9, 0x29, 0xEB, 0x2B, 0x2A, 0xEA, 0xEE, 0x2E, 0x2F, 0xEF, 0x2D, 0xED, 0xEC, 0x2C, 0xE4, 0x24, 0x25, 0xE5,
-    0x27, 0xE7, 0xE6, 0x26, 0x22, 0xE2, 0xE3, 0x23, 0xE1, 0x21, 0x20, 0xE0, 0xA0, 0x60, 0x61, 0xA1, 0x63, 0xA3, 0xA2, 0x62, 0x66, 0xA6, 0xA7, 0x67, 0xA5, 0x65, 0x64, 0xA4, 0x6C,
-    0xAC, 0xAD, 0x6D, 0xAF, 0x6F, 0x6E, 0xAE, 0xAA, 0x6A, 0x6B, 0xAB, 0x69, 0xA9, 0xA8, 0x68, 0x78, 0xB8, 0xB9, 0x79, 0xBB, 0x7B, 0x7A, 0xBA, 0xBE, 0x7E, 0x7F, 0xBF, 0x7D, 0xBD,
-    0xBC, 0x7C, 0xB4, 0x74, 0x75, 0xB5, 0x77, 0xB7, 0xB6, 0x76, 0x72, 0xB2, 0xB3, 0x73, 0xB1, 0x71, 0x70, 0xB0, 0x50, 0x90, 0x91, 0x51, 0x93, 0x53, 0x52, 0x92, 0x96, 0x56, 0x57,
-    0x97, 0x55, 0x95, 0x94, 0x54, 0x9C, 0x5C, 0x5D, 0x9D, 0x5F, 0x9F, 0x9E, 0x5E, 0x5A, 0x9A, 0x9B, 0x5B, 0x99, 0x59, 0x58, 0x98, 0x88, 0x48, 0x49, 0x89, 0x4B, 0x8B, 0x8A, 0x4A,
-    0x4E, 0x8E, 0x8F, 0x4F, 0x8D, 0x4D, 0x4C, 0x8C, 0x44, 0x84, 0x85, 0x45, 0x87, 0x47, 0x46, 0x86, 0x82, 0x42, 0x43, 0x83, 0x41, 0x81, 0x80, 0x40,
-};
-
-uint16_t compute_crc(uint8_t* buf, uint16_t len)
-{
-    uint16_t hi = 0xFF;
-    uint16_t lo = 0xFF;
-    uint16_t i = 0;
-
-    while (len--) {
-        i = lo ^ *(buf++);
-        lo = (uint16_t)(hi ^ crc_hi[i]);
-        hi = crc_lo[i];
-    }
-    return (lo << 8 | hi);
-}
-
-static void serial_modbus_task_func(void* param)
-{
-    uint8_t buf[BUF_SIZE];
-    uart_event_t event;
-
-    while (true) {
-        if (xQueueReceive(uart_queue, (void*)&event, portMAX_DELAY)) {
-            switch (event.type) {
-            case UART_DATA:
-                if (event.timeout_flag) {
-                    uint16_t len = uart_read_bytes(port, buf, event.size, portMAX_DELAY);
-
-                    ESP_LOG_LEVEL(LOG_LVL_DATA, TAG, "Received buffer length %d", event.size);
-                    ESP_LOG_BUFFER_HEX_LEVEL(TAG, buf, event.size, LOG_LVL_DATA);
-
-                    len -= 2;
-                    if (modbus_filter_request(buf, len)) {
-                        if (compute_crc(buf, len) == MODBUS_READ_UINT16(buf, len)) {
-                            len = modbus_request_exec(buf, len);
-                            if (len > 0) {
-                                MODBUS_WRITE_UINT16(buf, len, compute_crc(buf, len));
-                                len += 2;
-
-                                ESP_LOG_LEVEL(LOG_LVL_DATA, TAG, "Write buffer length %d", len);
-                                ESP_LOG_BUFFER_HEX_LEVEL(TAG, buf, len, LOG_LVL_DATA);
-
-                                uart_write_bytes(port, buf, len);
-                            }
-                        } else {
-                            ESP_LOGW(TAG, "Invalid packet CRC");
-                        }
-                    }
-                }
-                break;
-            case UART_FIFO_OVF:
-                xQueueReset(uart_queue);
-                break;
-            case UART_BUFFER_FULL:
-                xQueueReset(uart_queue);
-                uart_flush_input(port);
-                break;
-            default:
-                break;
-            }
-        }
-    }
-}
-
-void serial_modbus_start(uart_port_t uart_num, uint32_t baud_rate, uart_word_length_t data_bits, uart_stop_bits_t stop_bit, uart_parity_t parity, bool rs485)
-{
-    ESP_LOGI(TAG, "Starting on uart %d", uart_num);
-
-    uart_config_t uart_config = {
-        .baud_rate = baud_rate,
-        .data_bits = data_bits,
-        .parity = parity,
-        .stop_bits = stop_bit,
-        .flow_ctrl = UART_HW_FLOWCTRL_DISABLE,
-        .rx_flow_ctrl_thresh = 2,
-        .source_clk = UART_SCLK_DEFAULT,
-    };
-
-    esp_err_t err = uart_param_config(uart_num, &uart_config);
-    if (err != ESP_OK) {
-        ESP_LOGE(TAG, "uart_param_config() returned 0x%x", err);
-        return;
-    }
-
-    err = uart_driver_install(uart_num, BUF_SIZE, BUF_SIZE, EVENT_QUEUE_SIZE, &uart_queue, ESP_INTR_FLAG_LOWMED);
-    if (err != ESP_OK) {
-        ESP_LOGE(TAG, "uart_driver_install() returned 0x%x", err);
-        return;
-    }
-    port = uart_num;
-
-    if (rs485) {
-        err = uart_set_mode(uart_num, UART_MODE_RS485_HALF_DUPLEX);
-        if (err != ESP_OK) {
-            ESP_LOGE(TAG, "uart_set_mode() returned 0x%x", err);
-            return;
-        }
-        err = uart_set_rx_timeout(uart_num, 3);
-        if (err != ESP_OK) {
-            ESP_LOGE(TAG, "uart_set_rx_timeout() returned 0x%x", err);
-            return;
-        }
-    }
-
-    uart_set_always_rx_timeout(uart_num, true);
-
-    xTaskCreate(serial_modbus_task_func, "srl_modbus", 2 * 1024, NULL, 5, &serial_modbus_task);
-}
-
-void serial_modbus_stop(void)
-{
-    ESP_LOGI(TAG, "Stopping");
-
-    if (serial_modbus_task) {
-        vTaskDelete(serial_modbus_task);
-        serial_modbus_task = NULL;
-    }
-
-    if (port != -1) {
-        esp_err_t err = uart_driver_delete(port);
-        if (err != ESP_OK) {
-            ESP_LOGE(TAG, "uart_driver_delete() returned 0x%x", err);
-        }
-        port = -1;
-    }
-}
