+++ dependencies.lock
@@ -1,4 +1,14 @@
 dependencies:
+  espressif/cjson:
+    component_hash: e788323270d90738662d66fffa910bfe1fba019bba087f01557e70c40485b469
+    dependencies:
+    - name: idf
+      require: private
+      version: '>=5.0'
+    source:
+      registry_url: https://components.espressif.com/
+      type: service
+    version: 1.7.19~2
   espressif/mdns:
     component_hash: 3ec0af5f6bce310512e90f482388d21cc7c0e99668172d2f895356165fc6f7c5
     dependencies:
@@ -9,10 +19,20 @@ dependencies:
       registry_url: https://components.espressif.com/
       type: service
     version: 1.8.2
+  espressif/mqtt:
+    component_hash: ffdad5659706b4dc14bc63f8eb73ef765efa015bf7e9adf71c813d52a2dc9342
+    dependencies:
+    - name: idf
+      require: private
+      version: '>=5.3'
+    source:
+      registry_url: https://components.espressif.com/
+      type: service
+    version: 1.0.0
   idf:
     source:
       type: idf
-    version: 5.5.2
+    version: 6.0.0
   joltwallet/littlefs:
     component_hash: 362f1f5beb5087b0c60169aff82676d2d0ffc991ead975212b0cba95959181c5
     dependencies:
@@ -24,8 +44,10 @@ dependencies:
       type: service
     version: 1.14.8
 direct_dependencies:
+- espressif/cjson
 - espressif/mdns
+- espressif/mqtt
 - joltwallet/littlefs
-manifest_hash: 6669ac5ade2969c12a691a52c6d7b8da12a6c79df4cfc8c8fefdc306ba474f4a
+manifest_hash: 352b734e078c2c633c2d5f32745752c8cffc9de43d13bd9be7d0efaae18eae32
 target: esp32
-version: 2.0.0
+version: 3.0.0
